chore: fixed requested reviews
